+++ app/db/init.py
@@ -1,0 +1,10 @@
+"""Initialize the database."""
+
+from app.db.engine import engine
+from app.models.base_orm import Base
+
+
+async def init_db() -> None:
+    """Initialize the database by creating all tables."""
+    async with engine.begin() as conn:
+        await conn.run_sync(Base.metadata.create_all)
